ayu: Save 8 bytes on most accessors by unioning type with access function, and using a switch on a u8 instead of a virtual function for most accessor forms.
It just so happens that all the accessor forms that need an access function don't need a type, and vice versa.  This isn't a coincidence, because the reason for needing an access function is to preserve type information in the code, and if that information is in the code, it doesn't need to be stored in the data.